+++ src/components/pages/apps/genes/gsea/gsea-plot/gsea-plot-page.tsx
@@ -459,7 +459,6 @@ export function GseaPlotQueryPage() {
     <ClientLayout>
       <AxesPlotProvider>
         <GseaPlotProvider>
-          {' '}
           <GseaPlotPage />
         </GseaPlotProvider>
       </AxesPlotProvider>