+++ src/components/plot/axes/plot/axis-plot-props-popover.tsx
@@ -6,18 +6,17 @@ import {
 import { ToolbarIconButton } from '@/components/toolbar/toolbar-icon-button'
 import { Move3d } from 'lucide-react'
 import { useState } from 'react'
+import { IPlotAddress } from '../axes-provider'
 import { AxisPlotPropsPanel } from './axis-plot-props-panel'
 
 export function AxisPlotPropsPopover({
   title,
   axis,
-  plotId,
-  axisId,
+  plotAddress,
 }: {
   title: string
   axis: 'x' | 'y'
-  plotId: string
-  axisId: string
+  plotAddress: IPlotAddress
 }) {
   const [open, setOpen] = useState(false)
 
@@ -41,7 +40,7 @@ export function AxisPlotPropsPopover({
       />
 
       <PopoverContent className="gap-y-1 w-60">
-        <AxisPlotPropsPanel title={title} plotId={plotId} axisId={axisId} />
+        <AxisPlotPropsPanel title={title} plotAddress={plotAddress} />
       </PopoverContent>
     </Popover>
   )