+++ src/components/plot/axes/plot/tick-plot-props-popover.tsx
@@ -14,26 +14,26 @@ import { capitalCase } from '@/lib/text/capital-case'
 import { produce } from 'immer'
 import { MoveRight, MoveUp } from 'lucide-react'
 import { useState } from 'react'
-import { useAxes } from '../axes-provider'
+import { IPlotAddress, useAxes } from '../axes-provider'
 import { getAxisFormatter, getAxisTicks } from '../axis'
 
 export function TickPlotPropsPopover({
   title,
-  plotId,
-  axisId,
+  plotAddress,
   which,
 }: {
   title: string
-  plotId: string
-  axisId: string
+  plotAddress: IPlotAddress
   which: 'major' | 'minor'
 }) {
   const [open, setOpen] = useState(false)
 
   const { plots, updateAxis } = useAxes()
 
+  const { plotId, groupId, axisId } = plotAddress
   const plot = plots[plotId]
-  const axis = plot.axes[axisId]
+  const group = plot.groups[groupId]
+  const axis = group.axes[axisId]
   const ticks = axis.ticks[which]
 
   const items = getAxisTicks(axis, { which })
@@ -59,7 +59,7 @@ export function TickPlotPropsPopover({
           className="font-bold"
           checked={ticks.show}
           onCheckedChange={(v) => {
-            updateAxis(plotId, axisId, {
+            updateAxis(plotAddress, {
               ticks: produce(axis.ticks, (draft) => {
                 draft[which].show = v
               }),
@@ -73,7 +73,7 @@ export function TickPlotPropsPopover({
                 textProps: ticks.style.labels,
                 showRotation: true,
                 update: (f) =>
-                  updateAxis(plotId, axisId, {
+                  updateAxis(plotAddress, {
                     ticks: produce(axis.ticks, (draft) => {
                       draft[which].style.labels = Object.assign(
                         {},
@@ -92,7 +92,7 @@ export function TickPlotPropsPopover({
           tooltip="Specify the tick values using semicolons (e.g., 1; 2; 3) or use 'auto' for automatic ticks."
           checked={axis.ticks[which].style.line.show}
           onCheckedChange={(v) => {
-            updateAxis(plotId, axisId, {
+            updateAxis(plotAddress, {
               ticks: produce(axis.ticks, (draft) => {
                 draft[which].style.line.show = v
               }),
@@ -103,7 +103,7 @@ export function TickPlotPropsPopover({
             value={items.map((v) => format(v.v)).join('; ')}
             onTextChanged={(v) => {
               if (v === 'auto') {
-                updateAxis(plotId, axisId, {
+                updateAxis(plotAddress, {
                   ticks: produce(axis.ticks, (draft) => {
                     draft[which].items = undefined
                   }),
@@ -115,7 +115,7 @@ export function TickPlotPropsPopover({
                 .split(';')
                 .map((s) => parseFloat(s.trim().replace(/,/g, '')))
                 .filter(Number.isFinite)
-              updateAxis(plotId, axisId, {
+              updateAxis(plotAddress, {
                 ticks: produce(axis.ticks, (draft) => {
                   draft[which].items = values.map((v) => ({
                     v,
@@ -133,7 +133,7 @@ export function TickPlotPropsPopover({
           tooltip="Specify the tick labels using semicolons (e.g., 1; 2; 3)"
           checked={axis.ticks[which].style.labels.show}
           onCheckedChange={(v) => {
-            updateAxis(plotId, axisId, {
+            updateAxis(plotAddress, {
               ticks: produce(axis.ticks, (draft) => {
                 draft[which].style.labels.show = v
               }),
@@ -166,7 +166,7 @@ export function TickPlotPropsPopover({
                 label: i < values.length ? values[i] : '',
               }))
 
-              updateAxis(plotId, axisId, {
+              updateAxis(plotAddress, {
                 ticks: produce(axis.ticks, (draft) => {
                   draft[which].items = newItems
                 }),
@@ -185,7 +185,7 @@ export function TickPlotPropsPopover({
             limit={[1, 1000]}
             dp={0}
             onNumChanged={(v) => {
-              updateAxis(plotId, axisId, {
+              updateAxis(plotAddress, {
                 ticks: produce(axis.ticks, (draft) => {
                   draft[which].style.line.size = v
                 }),
@@ -197,7 +197,7 @@ export function TickPlotPropsPopover({
             value={ticks.style.line.offset}
             title="Offset"
             onNumChanged={(v) => {
-              updateAxis(plotId, axisId, {
+              updateAxis(plotAddress, {
                 ticks: produce(axis.ticks, (draft) => {
                   draft[which].style.line.offset = v
                 }),
@@ -210,7 +210,7 @@ export function TickPlotPropsPopover({
             value={ticks.style.labels.offset}
             title="Label Offset"
             onNumChanged={(v) => {
-              updateAxis(plotId, axisId, {
+              updateAxis(plotAddress, {
                 ticks: produce(axis.ticks, (draft) => {
                   draft[which].style.labels.offset = v
                 }),